Refactor: API 응답 DTO의 ID 타입을 Long으로 통일 및 userId 포함

See merge request s13-webmobile4-sub1/S13P11D104!46

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/chat/dto/ChatMessageDTO.java
@@ -13,8 +13,8 @@ import java.util.List;
 @AllArgsConstructor
 public class ChatMessageDTO {
     private Long chatId;
-    private Integer chatRoomId;
-    private Integer userId;
+    private Long chatRoomId;
+    private Long userId;
     private String nickname;
     private String message;
     private List<String> imageUrls; // 여러 이미지 URL
